Updated V001 to accept API key in args Added the same to V002 Put pieces together in v002

+++ src/Jarvis_0_0_2/AI_STT.py
@@ -7,23 +7,32 @@ import pyaudio
 import math
 import wave
 import tempfile
+from AI_Call import AIConnect
 
 class SpeechToText:
 	def __init__(self):
 		self.thread = None
+		#Model Variables
 		self.model = pywhisper.load_model("base.en")
 		self.temp_file = tempfile.NamedTemporaryFile(suffix=".wav", delete_on_close=False)
 		self.running = True
+		#Audio Variables
 		self.audio_threshold = 1000
 		self.sample_rate = 16000
 		self.bits_per_sample = 16
 		self.chunk_size = 1024
 		self.audio_format = pyaudio.paInt16
 		self.channels = 1
+		#Flag for determining if the audio can be read or not
+		self.flag = False
 
 	def start(self):
 		self.thread = Thread(target=self._talk, name="STT_Microphone_Input")
 		self.thread.start()
+
+	def stop(self):
+		self.running = False
+		self.thread.join()
 
 	def _detect_input(self, data) -> bool:
 		if not data:
@@ -33,6 +42,8 @@ class SpeechToText:
 
 	def _talk(self):
 		while self.running:
+			if self.flag: continue
+
 			self._setup_audio()
 
 			chunks_checked = 0
@@ -41,7 +52,7 @@ class SpeechToText:
 				data = self.stream.read(self.chunk_size)
 				chunks_checked += 1
 
-			self._read_audio()
+			self.flag = True
 
 	def _setup_audio(self):
 		self.wav_file = wave.open(f=self.temp_file.file, mode="wb")
@@ -64,13 +75,22 @@ class SpeechToText:
 			stream_callback=callback
 		)
 
-	def _read_audio(self):
+	def read_audio(self):
+		if not self.flag: return None
+		#Stopping All Sources so we can read the file
 		self.stream.stop_stream()
 		self.stream.close()
 		self.audio.terminate()
 		self.wav_file.close()
 		self.temp_file.close()
+		#Reading/Transcribing the file
 		result = pywhisper.transcribe(model=self.model, audio=self.temp_file.name)
+		#Removing the file
 		os.remove(self.temp_file.name)
+		#Creating a new temp_file
 		self.temp_file = tempfile.NamedTemporaryFile(suffix=".wav", delete_on_close=False)
+		self.flag = False
 		return result["text"].strip()
+
+	def is_running(self) -> bool:
+		return self.running
